move segment to modules folder
